Add files via upload
# Kickstarting with Excel
## Overview of Project
Analyze the success of campaign outcomes based on established fundraising goals and launch dates
### Purpose 
The purpose of the project is to determine the outcome of Kickstarter campaigns based on when the campaign is launched and the fund raising goals to help Louise decide when is the best time to launch her campaign and the ideal budget range for the project. 
## Analysis and Challenges
### Analysis of Outcomes Based on Launched Date
I looked at the percentage of successful, failed, and canceled projects and the goals that were set for the campaigns. Based on historical data, Louise has the best chance of success if she launches her Kickstarter campaign in May, when there is a peak in the data. The number of successful campaigns begin to taper off from June to August. It is not recommended to launch a campaign in December as this has the least successful campaigns. Refer to Outcomes vs Launch chart. 
### Analysis of Outcomes Based on Goals
Most of the fund-raising goals are less that $4999. The campaigns with funding raising goals of less than $1000 also tend to have the most successful outcomes, followed by goals between $1000 and $4999. As the goal increases the number of successful campaigns quickly diminishes. Refer to Outcomes vs Goals chart.
### Challenges and Difficulties Encountered
One of the challenges I encountered was creating the correct formula to obtain the number of successful, failed, and canceled campaigns, using the goal ranges. I tried creating a couple variations of the formulas before I was successful. 
## Results
-What are two conclusions you can draw about the Outcomes based on Launch Date?
Louise should aim to launch her Kickstarter campaign in May, as she is most likely to have a successful campaign. She should avoid launching her campaign in the month of December. 
-What can you conclude about the Outcome based on Goals?
Louise should try to keep her goal below $4999 if she wants to increase the likelihood of success. 
-What are some limitations of this dataset?
Some of the limitations of the dataset is the timeframe, in which it was gathered. The trends in fundraising could have changed. It would be nice to know the method that were used to raise funds. For example, were the campaigns posted on social media, was there a mass email sent out, etc. 
-What are some other possible tables and/or graphs that we could create?
Some of tables that may be helpful, would be to look at the number of backers and how much each person contributed.